+++ src/testUtils/veraid/authority/MockAuthorityClient.ts
